// apps/api/src/db/generated/hire_me/Role.ts
// @generated
// This file is automatically generated by Kanel. Do not modify manually.

import type { ColumnType, Insertable, Selectable, Updateable } from "kysely";
import { IPostgresInterval } from "postgres-interval";
import type { CompanyId } from "./Company";
import type { default as ContractType } from "./ContractType";

/** Identifier type for hire_me.role */
export type RoleId = number & { __brand: "RoleId" };

/** Represents the table hire_me.role */
export default interface RoleTable {
	id: ColumnType<RoleId, never, never>;

	company_id: ColumnType<CompanyId, CompanyId, CompanyId>;

	title: ColumnType<string, string, string>;

	notes: ColumnType<string | null, string | null, string | null>;

	ad_url: ColumnType<string | null, string | null, string | null>;

	type: ColumnType<ContractType, ContractType, ContractType>;

	term: ColumnType<
		IPostgresInterval | null,
		IPostgresInterval | null,
		IPostgresInterval | null
	>;

	date_added: ColumnType<Date, Date | string | undefined, Date | string>;
}

export type Role = Selectable<RoleTable>;

export type NewRole = Insertable<RoleTable>;

export type RoleUpdate = Updateable<RoleTable>;